add unit tests

+++ backend/Alpaki.Logic/Handlers/DeleteUser/DeleteUserHandler.cs
@@ -1,7 +1,7 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Alpaki.CrossCutting.Interfaces;
 using Alpaki.Database;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
@@ -10,24 +10,28 @@ namespace Alpaki.Logic.Handlers.DeleteUser
 {
     public class DeleteUserHandler : IRequestHandler<DeleteUserRequest, DeleteUserResponse>
     {
-        private readonly ICurrentUserService _currentUserService;
         private readonly IDatabaseContext _databaseContext;
 
-        public DeleteUserHandler(ICurrentUserService currentUserService, IDatabaseContext databaseContext)
+        public DeleteUserHandler(IDatabaseContext databaseContext)
         {
-            _currentUserService = currentUserService;
             _databaseContext = databaseContext;
         }
 
         public async Task<DeleteUserResponse> Handle(DeleteUserRequest request, CancellationToken cancellationToken)
         {
             var userToRemove = _databaseContext.Users.Where(u => u.UserId == request.UserId);
+
+            if (!(await userToRemove.AnyAsync()))
+            {
+                throw new ArgumentException($"User with given UserId does not exists [UserId={request.UserId}");
+            }
+
             var assignedDreamsConnections = _databaseContext.AssignedDreams.Where(a => a.VolunteerId == request.UserId);
 
             _databaseContext.AssignedDreams.RemoveRange(assignedDreamsConnections);
             _databaseContext.Users.RemoveRange(userToRemove);
 
-            await _databaseContext.SaveChangesAsync();
+            await _databaseContext.SaveChangesAsync(cancellationToken: cancellationToken);
 
             return new DeleteUserResponse();
         }
